refactor: move userInfo component location
